+++ src/main/CharacterSets.ts
@@ -16,6 +16,7 @@ export class CharacterSet {
   CONTRACTION_30: string;
   CONTRACTION_40: string;
   CONTRACTION_50: string;
+  CONTRACTION_60: string;
   useTraditional: boolean;
   useCapital: boolean;
 
@@ -37,14 +38,15 @@ export class CharacterSet {
     this.CONTRACTION_30 = contraction[1];
     this.CONTRACTION_40 = contraction[2];
     this.CONTRACTION_50 = contraction[3];
+    this.CONTRACTION_60 = contraction[4];
   }
 }
 
 const allCharacterSets = [
   //
-  new CharacterSet('零一二三四五六七八九', '十百千万亿兆', '正负', '点', '廿卅卌圩', false, false),
+  new CharacterSet('零一二三四五六七八九', '十百千万亿兆', '正负', '点', '廿卅卌圩圆', false, false),
   new CharacterSet('零壹贰叁肆伍陆柒捌玖', '拾佰仟萬億兆', '正负', '点', '念', false, true),
-  new CharacterSet('零一二三四五六七八九', '十百千萬億兆', '正負', '點', '廿卅卌圩', true, false),
+  new CharacterSet('零一二三四五六七八九', '十百千萬億兆', '正負', '點', '廿卅卌圩圆', true, false),
   new CharacterSet('零壹貳參肆伍陸柒捌玖', '拾佰仟萬億兆', '正負', '點', '念', true, true),
 ];